backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-multipart>=0.0.18
httpx>=0.28.0
websockets>=14.0
python-socks>=2.0.0
webrtcvad-wheels>=2.0.10
openai>=1.60.0
pymupdf>=1.25.0
python-docx>=1.1.0
chromadb>=0.6.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.20.0
alembic>=1.14.0
requests>=2.32.0
pyyaml>=6.0.0
jinja2>=3.1.0
